legacy-code-alpha1/app/requirements.txt
streamlit>=1.33.0 # Für die Web-Oberfläche
supabase>=2.0.0   # Python-Client für Supabase
python-dotenv>=1.0.0 # Zum Laden von Umgebungsvariablen aus .env
dspy-ai==3.0.3  # DSPy 3.0.3 gepinnt für schnelleren Build und Reproduzierbarkeit
# dspy==2.5.43     # Removed - use latest 3.x version
cloudpickle      # Required by DSPy for module serialization
openai>=1.68.2  # Updated für DSPy 3.0.3 LiteLLM Kompatibilität
requests          # Oft nützlich für API-Aufrufe (DSPy benötigt es implizit oft)
pandas>=1.5.0    # Falls wir Daten tabellarisch verarbeiten/anzeigen wollen
ollama
python-dateutil>=2.8.0  # Für robustes Datum/Zeit-Parsing
plotly>=5.14.0  # Für interaktive Visualisierungen (Fortschrittsbalken)

# Für Datei-Upload und Vision-Processing:
Pillow>=9.0.0    # Für Bildverarbeitung und Normalisierung
PyMuPDF>=1.23.0  # Für PDF zu Image Konvertierung

# Security dependencies:
bleach>=6.0.0  # Für HTML-Sanitization gegen XSS

# Test dependencies:
colorama>=0.4.6  # Für farbige Konsolen-Ausgabe in Tests

# Cookie Management
extra-streamlit-components>=0.1.60